requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
cffi==1.15.1
colorama==0.4.6
cryptography==39.0.1
decorator==5.1.1
Django==4.1.7
django-filter==1.1.0
djangorestframework==3.14.0
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==5.2.2
executing==1.2.0
ipdb==0.13.11
ipython==8.10.0
jedi==0.18.2
matplotlib-inline==0.1.6
numpy==1.24.2
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.37
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
tzdata==2022.7
wcwidth==0.2.6
